Use a rust executable in tests
Fixes #10.

- I decided to introduce another folder (tests/testees) to keep
programs under test separated. This simplifies the
Makefile because we don't need to filter out the test code files like
readmem.rs

- I'm using ```--dynamic-no-pic``` to disable PIE (as suggested in #10).
According to the docs it should be used for *relocatable external
references, non-relocatable code* [1]. This guarantees that
executable symbols addresses are absolute and can be used for
referencing the symbols at runtime. I'm not sure what it's meant
by *relocatable external references*, but I'm assuming it's relate
with library references, and not relevant for now.

[1] -  https://doc.rust-lang.org/rustc/codegen-options/index.html