Use generic function to split memory operations
We can't use proc_vm_write and proc_vm_read on protected memory pages,
so we should split memory operations into the ones that can be applied to protected pages (using ptrace)
and the ones which can use more efficient operations (proc_vm_*). This commit changes read-specific function
to a generic function that can be used for writes too.