Fix writemem test failing with restricted ptrace mode
If ptrace_scope is set to 1, it runs in a restricted mode which makes
it more complicated to attach to a parent process from the child.
Instead, this commit reverses the debuggee/debugger roles and makes the
case more streamlined - the test result is inferred from the child process
exit code (for some reason, failed/panicking asserts don't set an exit code
explicitly). This also simplifies synchronisation and makes pipes redundant.