Bump gimli version to 24.0
this commit update the gimli and addr2line dependency

The load method was changed in version 24.0

Before, there was one load method which could be used to load the main object file as well as supplementary object file.
In the 24.0 version, there are two methods, load and load_sup. One for the main object file and the other for supplementary objects files.

since head crab does not currently support supplementary object files, this commit delete the stub loader which was use.